Rename I2C Drivers Folder (#533)
* Moved to Devices folder

* Moved to Devices folder

* Moved to Devices folder

* Moved to Devices folder

* Moved to Devices folder

* Moved to Devices folder

* Moved to Devices folder

* Updated namespace to Devices

* Updated namespace to Devices

* Updated namespace to Devices

* Updated namespace to Devices

* Revert "Updated namespace to Devices"

This reverts commit 3044f1add3b02b814777cdb7542dbeed76efd915.

* Revert "Updated namespace to Devices"

This reverts commit 9f23b31f033f3cd6934d202516f3c9d0f382d837.

* Revert "Updated namespace to Devices"

This reverts commit 3816d7ad1e7072e48c5b4ed52041dc434d3b7eb4.

* Revert "Revert "Updated namespace to Devices""

This reverts commit 3d8394f074cd192748454caf8821100ae67de3bc.

* Added using
