+++ src/System.Device.Gpio/System/Device/I2c/Devices/I2cDevice.Windows.cs
@@ -2,15 +2,15 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace System.Device.I2c
+namespace System.Device.I2c.Devices
 {
     public partial class I2cDevice
     {
         /// <summary>
-        /// Creates a communications channel to a device on an I2C bus running on Unix.
+        /// Creates a communications channel to a device on an I2C bus running on Windows 10 IoT.
         /// </summary>
         /// <param name="settings">The connection settings of a device on an I2C bus.</param>
-        /// <returns>A communications channel to a device on an I2C bus running on Unix.</returns>
-        public static I2cDevice Create(I2cConnectionSettings settings) => new Drivers.UnixI2cDevice(settings);
+        /// <returns>A communications channel to a device on an I2C bus running on Windows 10 IoT.</returns>
+        public static I2cDevice Create(I2cConnectionSettings settings) => new Windows10I2cDevice(settings);
     }
 }
