+++ src/System.Device.Gpio/System/Device/I2c/Devices/I2cDevice.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace System.Device.I2c
+namespace System.Device.I2c.Devices
 {
     /// <summary>
     /// The communications channel to a device on an I2C bus.
