+++ tools/DevicesApiTester/Commands/I2c/I2cCommand.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Device.I2c;
+using System.Device.I2c.Devices;
 using CommandLine;
 using DeviceApiTester.Infrastructure;
