+++ tools/DevicesApiTester/Commands/I2c/I2cDriverType.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Device.I2c.Drivers;
+using System.Device.I2c.Devices;
 using DeviceApiTester.Infrastructure;
 
 namespace DeviceApiTester.Commands.I2c
