+++ tools/DevicesApiTester/Commands/I2c/I2cDump.cs
@@ -4,6 +4,7 @@
 
 using System;
 using System.Device.I2c;
+using System.Device.I2c.Devices;
 using System.Text;
 using CommandLine;
 using DeviceApiTester.Infrastructure;
